fix int64_t warnning on 32bit machine #27

+++ lib/json.c
@@ -1335,7 +1335,7 @@ json_unescape (const char *text)
 					}
 					else
 					{
-						fprintf (stderr, "JSON: unsupported unicode value: 0x%lX\n", unicode);
+						fprintf (stderr, "JSON: unsupported unicode value: 0x%lX\n", (unsigned long)unicode);
 					}
 				}
 				break;
